feat: add new API endpoints and reporting features
- Implemented PDF report generation for slowest steps and bottleneck analysis in `pdf-perf-sections.js` and `pdf-sections.js`.
- Created performance table utilities in `perf-tables.js` for summarizing step durations and bottlenecks.
- Added new SonarCloud API functions for managing hotspots, issues, permissions, project configurations, quality gates, and quality profiles.
- Introduced utility functions for extracting rule impacts and severity mappings in `rule-helpers.js`.
- Developed system information utilities to log and manage performance configurations in `system-info.js`.
