+++ src/reports/format-pdf-executive.js
@@ -1,19 +1,11 @@
-/**
- * Executive Summary PDF generator.
- * Produces a polished, concise PDF aimed at engineering leadership.
- */
-
-import { formatDuration, computeProjectStats, computeOverallStatus, computeTotalDurationMs, getNewCodePeriodSkippedProjects } from './shared.js';
+import { formatDuration, computeProjectStats, computeOverallStatus, computeTotalDurationMs } from './shared.js';
 import { generatePdfBuffer, pdfStyles } from './pdf-helpers.js';
+import { buildWarnings, buildActionItems, buildFailedProjects } from './pdf-exec-sections.js';
 
-/**
- * Generate the executive summary as a PDF buffer.
- */
 export async function generateExecutiveSummaryPdf(results) {
   const stats = computeProjectStats(results);
   const overallStatus = computeOverallStatus(stats);
   const content = [];
-
   content.push(...buildHeader(results));
   content.push(...buildStatusBanner(stats, overallStatus));
   content.push(...buildKeyMetrics(results, stats));
@@ -50,13 +42,11 @@ function buildHeader(results) {
   const date = results.startTime
     ? new Date(results.startTime).toLocaleDateString('en-US', { year: 'numeric', month: 'long', day: 'numeric' })
     : 'Unknown';
-
   const nodes = [
     { text: 'CloudVoyager Migration', style: 'title' },
     { text: 'Executive Summary', fontSize: 14, color: '#666666', margin: [0, 0, 0, 10] },
     { text: `Date: ${date}`, style: 'metadata' },
   ];
-
   if (durationMs != null) {
     nodes.push({ text: `Duration: ${formatDuration(durationMs)}`, style: 'metadata' });
   }
@@ -67,7 +57,6 @@ function buildHeader(results) {
   if (orgCount > 0) {
     nodes.push({ text: `Target Organizations: ${orgCount}`, style: 'metadata' });
   }
-
   return nodes;
 }
 
@@ -75,16 +64,13 @@ function buildStatusBanner(stats, overallStatus) {
   const successRate = stats.total > 0
     ? ((stats.succeeded / stats.total) * 100).toFixed(1)
     : '0.0';
-
   let bannerStyle = 'bannerSuccess';
   if (overallStatus === 'FAILED') bannerStyle = 'bannerFail';
   else if (overallStatus === 'PARTIAL SUCCESS') bannerStyle = 'bannerPartial';
-
   const nodes = [
     { canvas: [{ type: 'line', x1: 0, y1: 0, x2: 515, y2: 0, lineWidth: 1, lineColor: '#cccccc' }], margin: [0, 10, 0, 0] },
     { text: `Overall Status: ${overallStatus}`, style: bannerStyle },
   ];
-
   if (stats.total === 0) {
     nodes.push({ text: 'No projects were migrated.', style: 'bodyText' });
   } else if (stats.failed === 0 && stats.partial === 0) {
@@ -98,7 +84,6 @@ function buildStatusBanner(stats, overallStatus) {
       nodes.push({ text: `${stats.failed} project(s) failed entirely.`, style: 'bodyText' });
     }
   }
-
   return nodes;
 }
 
@@ -119,93 +104,8 @@ function buildKeyMetrics(results, stats) {
     ['Issues Synced', `${results.issueSyncStats.matched} matched, ${results.issueSyncStats.transitioned} transitioned`],
     ['Hotspots Synced', `${results.hotspotSyncStats.matched} matched, ${results.hotspotSyncStats.statusChanged} status changed`],
   ];
-
   return [
     { text: 'Key Metrics', style: 'heading' },
-    { table: { headerRows: 1, widths: [200, '*'], body }, layout: 'lightHorizontalLines' },
-  ];
-}
-
-function buildWarnings(results, stats) {
-  const keyWarnings = results.projectKeyWarnings || [];
-  const ncpSkipped = getNewCodePeriodSkippedProjects(results);
-
-  if (keyWarnings.length === 0 && ncpSkipped.length === 0 && stats.failed === 0 && stats.partial === 0) {
-    return [
-      { text: 'Warnings & Risks', style: 'heading' },
-      { text: 'No warnings or risks identified.', style: 'bodyText' },
-    ];
-  }
-
-  const nodes = [{ text: 'Warnings & Risks', style: 'heading' }];
-
-  if (keyWarnings.length > 0) {
-    nodes.push({ text: 'Project Key Conflicts', style: 'subheading' });
-    nodes.push({ text: `${keyWarnings.length} project(s) required a renamed key on SonarCloud due to global key conflicts. This may affect CI/CD pipeline configurations.`, style: 'bodyText' });
-  }
-
-  if (ncpSkipped.length > 0) {
-    nodes.push({ text: 'New Code Period Configuration', style: 'subheading' });
-    nodes.push({ text: `${ncpSkipped.length} project(s) have unsupported new code period types and require manual configuration in SonarCloud.`, style: 'bodyText' });
-  }
-
-  if (stats.failed > 0) {
-    nodes.push({ text: 'Failed Projects', style: 'subheading' });
-    nodes.push({ text: `${stats.failed} project(s) failed to migrate entirely. Review the detailed migration report for root cause analysis.`, style: 'bodyText' });
-  }
-
-  if (stats.partial > 0) {
-    nodes.push({ text: 'Partially Migrated Projects', style: 'subheading' });
-    nodes.push({ text: `${stats.partial} project(s) had one or more steps fail. These may need manual intervention.`, style: 'bodyText' });
-  }
-
-  return nodes;
-}
-
-function buildActionItems(results, stats) {
-  const keyWarnings = results.projectKeyWarnings || [];
-  const ncpSkipped = getNewCodePeriodSkippedProjects(results);
-  const items = [];
-
-  if (keyWarnings.length > 0) {
-    items.push(`Update CI/CD pipelines for renamed project keys (${keyWarnings.length} project(s))`);
-  }
-  if (ncpSkipped.length > 0) {
-    items.push(`Manually configure new code periods in SonarCloud (${ncpSkipped.length} project(s))`);
-  }
-  if (stats.failed > 0) {
-    items.push(`Investigate and retry failed project migrations (${stats.failed} project(s))`);
-  }
-  if (stats.partial > 0) {
-    items.push(`Review partially migrated projects and fix failed steps (${stats.partial} project(s))`);
-  }
-  items.push('Review quality profile rule gaps in quality-profiles/quality-profile-diff.json');
-  items.push('Verify project permissions in SonarCloud dashboard');
-
-  const nodes = [{ text: 'Action Items', style: 'heading' }];
-  items.forEach((item, i) => {
-    nodes.push({ text: `${i + 1}. ${item}`, style: 'actionItem' });
-  });
-  return nodes;
-}
-
-function buildFailedProjects(results) {
-  const failedProjects = results.projects.filter(p => p.status === 'failed');
-  if (failedProjects.length === 0) return [];
-
-  const body = [
-    [
-      { text: 'Project Key', style: 'tableHeader' },
-      { text: 'Failed Steps', style: 'tableHeader' },
-    ],
-  ];
-  for (const project of failedProjects) {
-    const failedSteps = project.steps.filter(s => s.status === 'failed');
-    body.push([project.projectKey, failedSteps.map(s => s.step).join(', ')]);
-  }
-
-  return [
-    { text: 'Failed Projects', style: 'heading' },
     { table: { headerRows: 1, widths: [200, '*'], body }, layout: 'lightHorizontalLines' },
   ];
 }
