+++ src/reports/format-performance.js
@@ -1,16 +1,8 @@
-/**
- * Performance report markdown formatter.
- * Shows timing breakdowns for each segment of the migration.
- */
-
 import { formatDuration, computeTotalDurationMs } from './shared.js';
+import { sumDurations, getStepDuration, getConfigDuration, formatSlowestSteps, formatBottleneckAnalysis } from './perf-tables.js';
 
-/**
- * Format results as a performance report in markdown.
- */
 export function formatPerformanceReport(results) {
   const sections = [];
-
   sections.push(formatHeader(results));
   sections.push(formatOverview(results));
   sections.push(formatServerSteps(results));
@@ -18,9 +10,7 @@ export function formatPerformanceReport(results) {
   sections.push(formatProjectBreakdown(results));
   sections.push(formatSlowestSteps(results));
   sections.push(formatBottleneckAnalysis(results));
-
   sections.push('---\n*Generated by CloudVoyager*\n');
-
   return sections.filter(Boolean).join('\n');
 }
 
@@ -42,7 +32,6 @@ function formatOverview(results) {
   const avgPerProject = projectCount > 0 && durationMs != null
     ? formatDuration(Math.round(durationMs / projectCount))
     : '—';
-
   const lines = [
     '## Overview\n',
     '| Metric | Value |',
@@ -52,38 +41,30 @@ function formatOverview(results) {
     `| Average Time per Project | ${avgPerProject} |`,
     `| Organizations | ${(results.orgResults || []).length} |`,
   ];
-
-  // Add server step total
   const serverTotal = sumDurations(results.serverSteps);
   if (serverTotal > 0) {
     lines.push(`| Server-Wide Extraction | ${formatDuration(serverTotal)} |`);
   }
-
-  // Add org migration totals
   for (const org of (results.orgResults || [])) {
     if (org.durationMs != null) {
       lines.push(`| Org: ${org.key} (total) | ${formatDuration(org.durationMs)} |`);
     }
   }
-
   lines.push('');
   return lines.join('\n');
 }
 
 function formatServerSteps(results) {
   if (results.serverSteps.length === 0) return null;
-
   const lines = [
     '## Server-Wide Extraction\n',
     '| Step | Duration | Status |',
     '|------|----------|--------|',
   ];
-
   for (const step of results.serverSteps) {
     const dur = step.durationMs != null ? formatDuration(step.durationMs) : '—';
     lines.push(`| ${step.step} | ${dur} | ${step.status} |`);
   }
-
   const total = sumDurations(results.serverSteps);
   lines.push(`| **Total** | **${formatDuration(total)}** | |`);
   lines.push('');
@@ -92,31 +73,23 @@ function formatServerSteps(results) {
 
 function formatOrgSteps(results) {
   if (!results.orgResults || results.orgResults.length === 0) return null;
-
   const sections = [];
   for (const org of results.orgResults) {
-    const lines = [
-      `## Organization: ${org.key}\n`,
-    ];
-
+    const lines = [`## Organization: ${org.key}\n`];
     if (org.durationMs != null) {
       lines.push(`**Total org migration time:** ${formatDuration(org.durationMs)}\n`);
     }
-
     if (org.steps && org.steps.length > 0) {
       lines.push('| Step | Duration | Status | Detail |');
       lines.push('|------|----------|--------|--------|');
-
       for (const step of org.steps) {
         const dur = step.durationMs != null ? formatDuration(step.durationMs) : '—';
         const detail = step.detail || step.error || '';
         lines.push(`| ${step.step} | ${dur} | ${step.status} | ${detail} |`);
       }
-
       const total = sumDurations(org.steps);
       lines.push(`| **Total (org steps)** | **${formatDuration(total)}** | | |`);
     }
-
     lines.push('');
     sections.push(lines.join('\n'));
   }
@@ -125,138 +98,21 @@ function formatOrgSteps(results) {
 
 function formatProjectBreakdown(results) {
   if (results.projects.length === 0) return null;
-
-  // Sort by duration (slowest first)
   const sorted = [...results.projects].sort((a, b) => (b.durationMs || 0) - (a.durationMs || 0));
-
   const lines = [
     '## Per-Project Breakdown\n',
     'Sorted by total duration (slowest first).\n',
     '| # | Project | Total | Report Upload | Issue Sync | Hotspot Sync | Config |',
     '|---|---------|-------|---------------|------------|--------------|--------|',
   ];
-
   sorted.forEach((project, i) => {
     const total = project.durationMs != null ? formatDuration(project.durationMs) : '—';
     const reportUpload = getStepDuration(project, 'Upload scanner report');
     const issueSync = getStepDuration(project, 'Sync issues');
     const hotspotSync = getStepDuration(project, 'Sync hotspots');
     const configDur = getConfigDuration(project);
-
     lines.push(`| ${i + 1} | \`${project.projectKey}\` | ${total} | ${reportUpload} | ${issueSync} | ${hotspotSync} | ${configDur} |`);
   });
-
   lines.push('');
   return lines.join('\n');
-}
-
-function formatSlowestSteps(results) {
-  // Collect all steps across all projects
-  const allSteps = [];
-
-  for (const project of results.projects) {
-    for (const step of project.steps) {
-      if (step.durationMs != null && step.durationMs > 0) {
-        allSteps.push({
-          projectKey: project.projectKey,
-          step: step.step,
-          durationMs: step.durationMs,
-          status: step.status,
-        });
-      }
-    }
-  }
-
-  if (allSteps.length === 0) return null;
-
-  // Sort by duration descending, take top 10
-  allSteps.sort((a, b) => b.durationMs - a.durationMs);
-  const top = allSteps.slice(0, 10);
-
-  const lines = [
-    '## Slowest Individual Steps (Top 10)\n',
-    '| # | Project | Step | Duration | Status |',
-    '|---|---------|------|----------|--------|',
-  ];
-
-  top.forEach((entry, i) => {
-    lines.push(`| ${i + 1} | \`${entry.projectKey}\` | ${entry.step} | ${formatDuration(entry.durationMs)} | ${entry.status} |`);
-  });
-
-  lines.push('');
-  return lines.join('\n');
-}
-
-function formatBottleneckAnalysis(results) {
-  const durationMs = computeTotalDurationMs(results);
-  if (durationMs == null || durationMs === 0) return null;
-
-  // Aggregate time by step type across all projects
-  const stepTypeTotals = new Map();
-
-  for (const project of results.projects) {
-    for (const step of project.steps) {
-      const dur = step.durationMs || 0;
-      stepTypeTotals.set(step.step, (stepTypeTotals.get(step.step) || 0) + dur);
-    }
-  }
-
-  // Add server steps
-  for (const step of results.serverSteps) {
-    const dur = step.durationMs || 0;
-    stepTypeTotals.set(`[Server] ${step.step}`, (stepTypeTotals.get(`[Server] ${step.step}`) || 0) + dur);
-  }
-
-  // Add org steps
-  for (const org of (results.orgResults || [])) {
-    for (const step of (org.steps || [])) {
-      const dur = step.durationMs || 0;
-      const key = `[Org] ${step.step}`;
-      stepTypeTotals.set(key, (stepTypeTotals.get(key) || 0) + dur);
-    }
-  }
-
-  if (stepTypeTotals.size === 0) return null;
-
-  // Sort by cumulative time
-  const sorted = [...stepTypeTotals.entries()].sort((a, b) => b[1] - a[1]);
-  const totalStepTime = sorted.reduce((sum, [, dur]) => sum + dur, 0);
-
-  const lines = [
-    '## Bottleneck Analysis\n',
-    'Cumulative time spent on each step type across all projects.\n',
-    '| Step Type | Cumulative Time | % of Step Time |',
-    '|-----------|-----------------|----------------|',
-  ];
-
-  for (const [stepName, dur] of sorted) {
-    if (dur === 0) continue;
-    const pct = totalStepTime > 0 ? ((dur / totalStepTime) * 100).toFixed(1) : '0.0';
-    lines.push(`| ${stepName} | ${formatDuration(dur)} | ${pct}% |`);
-  }
-
-  lines.push(`| **Total** | **${formatDuration(totalStepTime)}** | **100%** |`);
-  lines.push('');
-  return lines.join('\n');
-}
-
-// --- helpers ---
-
-function sumDurations(steps) {
-  return (steps || []).reduce((sum, s) => sum + (s.durationMs || 0), 0);
-}
-
-function getStepDuration(project, stepName) {
-  const step = project.steps.find(s => s.step === stepName);
-  if (!step || step.durationMs == null) return '—';
-  if (step.status === 'skipped') return 'skipped';
-  return formatDuration(step.durationMs);
-}
-
-function getConfigDuration(project) {
-  // Config steps are everything except the three main steps
-  const mainSteps = new Set(['Upload scanner report', 'Sync issues', 'Sync hotspots']);
-  const configSteps = project.steps.filter(s => !mainSteps.has(s.step));
-  const total = configSteps.reduce((sum, s) => sum + (s.durationMs || 0), 0);
-  return total > 0 ? formatDuration(total) : '—';
 }
