+++ src/sonarqube/api/issues-hotspots.js
@@ -1,0 +1,25 @@
+import logger from '../../utils/logger.js';
+
+export async function getIssues(getPaginated, projectKey, filters = {}) {
+  logger.info(`Fetching issues for project: ${projectKey}`);
+  const params = { componentKeys: projectKey, ...filters };
+  return await getPaginated('/api/issues/search', params, 'issues');
+}
+
+export async function getIssuesWithComments(getPaginated, projectKey, filters = {}) {
+  logger.info(`Fetching issues with comments for project: ${projectKey}`);
+  const params = { componentKeys: projectKey, additionalFields: 'comments', ...filters };
+  return await getPaginated('/api/issues/search', params, 'issues');
+}
+
+export async function getHotspots(getPaginated, projectKey, filters = {}) {
+  logger.info(`Fetching hotspots for project: ${projectKey}`);
+  const params = { projectKey, ...filters };
+  return await getPaginated('/api/hotspots/search', params, 'hotspots');
+}
+
+export async function getHotspotDetails(client, hotspotKey) {
+  logger.debug(`Fetching hotspot details: ${hotspotKey}`);
+  const response = await client.get('/api/hotspots/show', { params: { hotspot: hotspotKey } });
+  return response.data;
+}
